chore: move tokens from contexts folder
