+++ packages/react-components/react-tree/src/contexts/index.ts
@@ -1,3 +1,2 @@
 export * from './treeContext';
-export * from './tokens';
 export * from './treeItemContext';
